orderservice/src/main/java/com/faikturan/orderservice/controller/OrderController.java
package com.faikturan.orderservice.controller;

import com.faikturan.orderservice.dto.OrderRequest;
import com.faikturan.orderservice.service.OrderService;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;

import static org.springframework.http.HttpStatus.CREATED;

@Tag(name = "Order", description = "the Order Api")
@RestController
@RequestMapping(value = "api/v1/order")
@RequiredArgsConstructor
public class OrderController {
    private final OrderService orderService;

    @PostMapping
    @ResponseStatus(CREATED)
    public String placeOrder(@RequestBody OrderRequest orderRequest) {
        orderService.placeOrder(orderRequest);
        return "Order Placed Successfully";
    }
}
